GDA-4932 - Removed the BasicScanDataPointPipeline
The intent was that the Multithreaded version always be used from 8.24 onwards. If pipeline length of zero is requested a warning ins printed and pipeline with length 1 is created instead. (Note that this will have no internal queue, and the one point it can hold is the one being completed and broadcast).