CLIENT: new About dialog image for GDA RCP products
Has some space around the edge to leave a gap between the about dialog border
and the logo/text within the image.

The image is specified on the "Branding" tab in the beamline product (in the
beamline feature).

Clicking "Synchronize" in the "Overview" tab in the product then copies the
image filename into an org.eclipse.core.runtime.products extension in the
beamline plugin.