GDA-4285 several fixes to XMap and ites live views.
The status of XMap was based on the TFG which drives it. But outside of
scans, i.e. when the rates views might be running, its status should be
based on its controller. If not then the data can be read out too fast 
and sometimes too early.

Also fixed the XES Xmap view so that the correct values are now being
used for the ICR and FF. The columns have been relabelled so that the
meaning of the values is more explicit.

This fix also helps the XMap Monitor view so as the Single-element/ I0
column could be inaccurate for the above bug in Xmap.

Sped up both views and removed the plot for single element XMap
detectors.