GDA-5276 Fix bug in BatonRequestDialog that causes baton ownership to suddenly change
To reproduce:
 * user A permission level < user B permission level
 * B has baton
 * A requests baton
 * (2-minute timer starts in B's client)
 * B agrees to pass the baton to A (chooses "Yes")
 * A gets the baton
 * B takes the baton again
 * 2-minute timer expires, and B passes the baton to A

BatonRequestDialog schedules a job to run 2 minutes after the dialog is opened.
This line:

    currentOpenDialog.setOk(returnCode == OK);

sets 'ok' to:
 * true if the user said "Yes"
 * false if the user said "No"

The job checks 'ok' and if it's true (i.e. the user chose "Yes"), the dialog is
assumed to be open, is closed, and the finally block causes the baton to be
passed to the requesting user.

Of course if the user chose "Yes", the dialog is no longer open, and their
choice will already have resulted in the baton getting passed. The job shouldn't
pass it again when the 2-minute timer expires.

It makes more sense if we change the 'ok' field to 'open', which means we can
just do:

    currentOpenDialog.setOpen(false);

This stops the job closing the dialog (and passing the baton) when the user
has already chosen "Yes".