CORE: store the full name of each user, in addition to their username
Clients still register themselves with the GDA server using just a username.
The server looks up the corresponding full name for that username.

There are however a couple of methods in JythonServerFacade (getBatonHolder and
getMyDetails) that avoid a trip to the server if the current user's details are
returned. In these cases, the client looks up the full name.

Obviously this relies on the server and client(s) having a consistent view of
user details.

Currently, the full name is just a copy of the username.