XAS-24 remove duplicate extension contributions in the shared plugin.
The contributions must be at a per-beamline level as each beamline has
its own unique list of beans/editors which it uses. 

Duplicating the contributions can cause problems in the GUI as when
reading the extension registry it may (as is) assumed that there is only
on contribution per editor/bean/scan type.