BLXX-76 improve how Bragg angle of spectrometer is derived before a move has been made by XESBragg
Any movement of XESBragg will ensure all the XES spectrometer motor
positions are consistent. Before a move is made the Bragg is now derived
from the det_y and xtal_x motors to give a roughly correct result. This 
should reduce the chance of a collision or slow move when the XESBragg
object is moved for the first time after a GDA restart.