Allow ProcessingDetectorWrapper to be configured not to include 'path' in its data
This is to fix a problem found on MX beamlines where a detector returned a 2D
dataset, and a ProcessingDetectorWrapper treated the dataset as a path, and
included the dataset in the data written to an SRS file. The 2D dataset
appearing in the SRS file (in an abbreviated form) prevented the file being
loaded properly.