fix metadata writing so that the header does not grow after every scan
The original array was being held by reference and so when the Jython 
code added any extra metadata both the live and stored copies were
altered.

The fix is to make a new object and give that to the MetaDataConfig
instead of altering the original.  Then replace the original at the end.

Also have removed the duplicated methods trying to do the same thing.