GDA -Core - checkForInterrupts does not check ScanStatus by default
The Java property ScanBase.GDA_SCAN_CLEAR_INTERRUPT_AT_SCAN_END 
has been changed to true by default. As a result 
checkForInterrupts will always throw InterruptedException if the
ScanBase.interrupted is true. Also on discovering that
ScanBase.interrupted is true at the end of a scan ScanBase.interrupted
will be set to false and an InterruptedException will be thrown.