fix bug where black pixels would be displays in map, which also means missing lines in rbg file
This was due to an error in the code when selecting which detector channel to plot during the scan. I have removed quite a bit of code which was only to add entries into the log anyway.

This highlights the weakness in the design where a change in the UI options has an impact on the data written to file...

The two need decoupling.