BLXVB-159, BLXX-214 fix the way the 'pause after current repetition' works
Instead of pausing the Command Queue (which pauses the Command Queue after the current job), simply set the ScriptBase paused flag and wait until the users clears it.

There is a bug in the Command Queue as it should not pause itself after current task once the user presses resume, but this fix is a proper fix too and would not need changing in any fix to Command Queue.