XAS-145 refactoring of Xspress2System
New class Xspress2CurrentSettings to hold output options for the system.
This is shared by the detector object and the data object which converts the raw data into Nexus.

Also the Reading classes are no longer internal and are in their own sub-package.

Other new classes also moved into the sub-package. This has caused minor changes to the imports of other classes in this plugin.

Units tests pass.