I18 acceptance testing: testing faster ratser scans
I had to clean up how the source of a map scan plot ( a current scan or
an old data file) was held. This was a mess as both the UI and server
held some compoeting state. Now UI knows where the plotted data comes
from and responds to user clicks accordingly.