reinstate old 8.36 versions of Xspress and Vortex XML editors
but leave the new versions which can be used in views. Ideally its these new versions should be supported in the long term as they will supply both views and editors.

But for the moment we have to use the old 8.36 versions and the new ones are just too buggy and there is not enough resource to get them working.