add boolean property and accessor for dummy mode
MX beamlines WERE using 'LocalProperties.check("gda.dummy.mode")' to
determine whether GDA is using the dummy mode configuration. HERE we
introduce a specific method (and constant) which enables better code
navigation for developers and can be less verbose, as described below.

isDummyModeEnabled() can be called unqualified in Java with 'static':

import static gda.configuration.properties.LocalProperties.isDummyModeEnabled;

or in Jython with:

from gda.configuration.properties.LocalProperties import isDummyModeEnabled

If used consistently, these import statements and isDummyModeEnabled() calls
effectively flag that a class/script has some differing logic depending on
whether or not it is run in dummy mode, which may be considered a
warning that dummy and live mode configurations are out-of-sync.