Allow disconnection from telnet/SSH server using Ctrl-D when JLine is turned on
Summary: the experience is *MUCH* better if you turn JLine on :-)

With JLine off, Ctrl-D is treated as part of a Jython command (and generally
messes things up). You still have to disconnect the client program manually
(Ctrl-] "quit" for telnet, Enter ~ . for SSH).

Additionally, for non-JLine SSH, you still can't see the characters you type
on-screen, as before.

With JLine on, things are now much better - if Ctrl-D is pressed, the socket
or session is closed, so the client gets disconnected. (Previously it would
just appear to hang, because the server had stopped processing data from the
client.)