correctly control speed in I18 QEXAFS scans
by giving the speed in keV/s instead of mDeg/s as it is the compound virtual EPICS motor of energy we are driving, not the bragg motor.

Changing the Bragg motor speed, and then moving the 'energy' motor will not have any impact on the speed of movement. The energy motor should have both its speed changed and then be operated on.