XAS-172 after testing QexafsScan.java
I was going to write a unit test but there is so much duplication with XasScan.java that all was needed was some refactoring to properly inherit from XasScan. There is not much to QexafsScan now.  All the real work is done in the preparers and the ContinuousScan class.

JavaDoc written for ExafsScan to explain the architecture.