BLXVIII-154 changes to xspress3 to work in 2D maps
Data cannot be written in a single HDF file by EPICS as that prevents the
MCAs from being viewed during a map, so instead each line is written to a
separate file and each file is linked to the GDA Nexus.

This is not ideal and creates an undesirable structure in the Nexus file
but this is the only option until SWMR for HDF becomes available.

The MCAs are now visible in the Microfocus perpspective during data collection.