add channel enable / disable functionality to Xspress3 change xspress subfolder
The MCAs from Xspress3 are written to their own HDF file. These are placed
in a subfolder called xspress3. For 2D scans there will be a folder below
named afer the scan. The HDF files themselves have the naming convention:
<scanNumber>_<rownumber>.hdf

So for 2D:
/dls/iXX/data/20XX/cm-XXXX/xspress3/<scan_num>/<scan_num>_<row_num>.hdf

for 1D:
/dls/iXX/data/20XX/cm-XXXX/xspress3/<scan_num>_<row_num>.hdf


when modeoverride (I18) then display options for virtual scalers, not windows

Change-Id: Ieeadd4c741a1a7eaf536bcae81e152ca1802a9a3