BLXVI-149 (8.42) -JS.getCurrentScanInformation returns null if no scan
Refactoring for GDA-6005 (commit 7223170) caused a behaviour change to
JythonServer.getCurrentScanInformation, causing it to throw a
NullPointerException if had not yet been a scan. Restored previous
behaviour of returning null in this case.

I16 run into this due to their use of NXDetector (they use a wrapper
class that manually calls NXDetector.atScanStart etc in response to the
pos command).