BLXVI-156 (8.42) - ProcessingDetectorWrapper warns if waitForFile==True
AreaDetectors normally have their tifwriters wait for the file to arrive
on the file system.
But since ProcessingDetectorWrapper already does this, having the
tifwriter do this is redundant and reduces performance.
Performance was especially degraded on I16 for Pilatus detectors using a
PPU.

No effect on detectors that do not have a tifwriter.