It should be acceptable to have a scan in a paused state as if finishes.
As the user may have changed the scan into a paused state at the last minute. This should not be a problem for the whole scan completeing normally.  However the assertion error does prevent the Command Queue from working and means the whole GDA needs a restart.

if scan object in 'paused' state at end of scan, then set to 'running' before the ending the scan to finish normally

This updates parts of the UI which assume the normal sequence.

Change-Id: Ie23c1fe1a0d7152e398a34800d3f9fd04df360a7