BLXXJ-315: ensure air supply is on for air bearing motor when motion is requested
class to control air supply for air bearing motor or other device require air supply. It takes two independent PVs, one for set, the other for read. It's isBusy() return false until the read and set is the same, as it takes seconds for air supply to reach its required pressure in the device.

Task-Url: http://jira.diamond.ac.uk/browse/BLXXJ-315