BLX-85 (8.44) Replace call to .getDimensions() with .shape in tryToLoadDataset
This problem was hidden by a swallowed exception in DatasetProvider.py
(/uk.ac.gda.core/scripts/gdascripts/analysis/io/DatasetProvider.py) which made
it look like files couldn't be read, when actually, tryToLoadDataset() was
trying to use a function which is no longer available on a dataset.

Add logging of the swallowed exception to make it easier to debug any future
problems.

Also add logging of the soft error which only went to the console.

Finally remove the call to touch, which makes it even more difficult to
track down the underlying problem. Add a TODO to remove this code
entirely if there is no compelling reason to keep it (which should be
documented).

Change-Id: I6beef37556a5cd612b17aca2e59ccdbf789e7a50