fix to xspress2 to correctly read ROIs when using VirtualScalers settings but reading out the full MCA, not partial MCAs
This is how I18 works: as multiple elements are studied at once, we need to read out the whole MCA and extract multiple ROIs (unlike B18 which typically only reads one element at a time so there we can use the hardware window instead of GDA calculating the ROIs).

This was not working in the refactored version. In the refactoring it was assumed that we would always be using partial-mcas or virtual scalers but never reading out the whole MCAs.

Change-Id: I307be4e3ccc2107d664825716350ef5fb444fd6d