BLXV (8.44) Fix problem with slow startup of GDA client after server startup.
Due to the problem with ssh where it doesn't return, gda_servers_core
runs ssh in the background and the scripts are expected to wait for the 
OBJECT_SERVER_STARTUP_FILE on the client side. Unfortunately, due to
network filesystem, this can take some time, and can delay the startup
of the GDA client by over a minute.

With this update, gda_servers_core records the PID of the SSH command in
an SERVERS_SSH_PID environment variable on the client side, while delaying
the exit of the ssh command until it sees the OBJECT_SERVER_STARTUP_FILE
on the server side. This allows the client side scripts to see that the
server has started up by either the OBJECT_SERVER_STARTUP_FILE appearing
on the filesystem or by the ssh command completing.

Thus the startup will complete quickly if ssh exits correctly or slowly
but reliably if the ssh command doesn't exit correctly.

This change requires commit 9866343 in gda-diamond.git in order to work
properly.

Change-Id: I8c22fc5bf4366c00999bff8fae640c4323101545