Hack fix when populating nodes with dead external links as children.
Populating a node in our cache could fail if the node has an external
link to a non-existent file as a child.
Quick fix stops us adding that node to our cache for the moment, but
this can result in an invalid cache if that file appears later.

Should add a "lazy" node in this case instead.