GDA-6220 Remove references to MockJythonServerFacade
The property which is checked (gda.localObjectsOnly) is not referenced anywhere else in the codebase. This dependency is the only thing stopping MockJythonServerFacade from being moved to the test helpers plugin, and since it seems to be unused I'm removing the option to use it during RCP application startup. I've also tidied a little of the rest of the code which was rendered mostly unnecessary by this change.

Change-Id: Ib063fd65a6296bb17dcae09ee0ab93593d17de89