Move ScanObject back to gda-core to improve split package problem.
In order to move forward with the OSGI server model we need to address the problem of split packages in the codebase. This is because bundles in OSGI use their own individual classloaders and so cannot access classes from other bundles even if they are in the same package, unless an artifical backdoor method like eclipse-buddies is used. Because of this, anything that currently uses ScanObject (i.e. most scans) fails in OSGI mode since the class was moved in the past to the uk.ac.gda.api bundle from uk.ac.gda.core. 

This change replaces the base class in uk.ac.gda.api with an equivalent Interface and moves a minimal version of the abstract class back to gda.core. This will not affect the python scripts (as they only know about packages not bundles) but allows the classes derived from ScanObject to work in OSGI mode, as they are now in the bit of the gda.scan package the scripts can see. The specifics are:

ScanObject.java - move from gda.api back to gda.core.

IScanObject.java - add to gda.api in a new package to replace Scan Object and preserve the definition of the API in the gda.api bundle.

Implicit/ExplicitScanObject.java - make the implementations of the new interface methods public.

gda-api/MANIFEST.MF  - export the new package that contains the interface

In theory, the other members of gda.scan in uk.ac.gda.api should be moved to the new package in the future as the refactoring progresses.

Change-Id: I344bde489afa3d2bd0d5ba17fb026b5244099d01