Add gda.analysis to dependencies to allow RCPPlotter class to move.
RCPPlotter is the single class in gda.client which causes gda.epics (and hence all server builds) to depend upon gda.client and thus many othe client plugins. Moving it to the gda.analysis plugin proper breaks this dependency and reduces the splitting of the gda.analysis package which we need to do to have an OSGI system anyway. To facilitate this I am adding the dependency on the uk.ac.gda.analysis plugin in advance of moving the class so that doing so won't break the build.

Change-Id: I6484c55b69269be95d54e886a02edfce9c332afd