BLXV (8.44) Log entries from bash into the GDA_LOG_FILE once it is defined.
This makes debugging problems with the bash startup scripts much easier,
even when GDA_DEBUG is not defined, as long as the problem is after logging
is set up.

Also, set up logging slightly earlier, so that more bashlog messages are
written to the log file, and remove now duplicate messages in the log file.

Change-Id: Ic457b96f9f1f8e23b55f60dce654b279733d9a0c