Fix energy calculation
Previously the code called self.energy_scannable(), which implicitly calls
getPosition(). A recent change accidentally removed the brackets, meaning
the variable was assigned to be a scannable rather than just a number, and
failed when used in arithmetic a couple of lines below. To make the behaviour
more clear, we now call getPosition() explicitly.

Change-Id: I8635c494561716bdc45da18a0d5377f0741b031d