Fix RBAC by adding net.sf.cglib dependencies
Under OSGi the Spring proxying mechanism cannot find the net.sf.gclib classes
used when creating proxy objects used in the creation of interceptors. To get
round this, dependencies on the core and proxy packages need to be explicitly
added to the affected plugins as well as org.python.core.

Change-Id: I98c9bdb8c16f0a1d3bfbabcab95bdfcc854d0b08