Allow collection strategies & decorators to save & restore state
prepareForCollection(), the function that should be called from outside
the strategies/decorator hierarchy, now calls saveState() to save the 
detector state (e.g. number of images, trigger mode), followed by 
rawPrepareForCollection(), which implements the functionality that
was previously in prepareForCollection().

The implementation of saveState() & rawPrepareForCollection()
need not do anything, but the decorator versions should at least
call the corresponding function in the next decoratee in the hierarchy.
Default implementations are provided which do this.

Parallel changes have been made in gda-core & gda-epics

Change-Id: I3bc4c4375ee3485c662a30ca0f951fb41b760388