DummyMotor: save new position to file at end of move
In what appears to be rather large oversight in DummyMotor, the motor's new
position is not saved to file when the move is complete.

Add a savePosition call near the end of DummyMotor.run.

There are few places this call could go, but I've chosen to make the position
save part of the move, by doing it *before* moving/busy flags are cleared.

Change-Id: I89428c5de6bb780485d4c74c959ea123fe28e9bd