Fix ScriptDrivenScannableMotionUnitsTest.testAsynchronousMoveTo
Up to now a DummyMotor hasn't actually saved its position to file, as it's
supposed to. In ScriptDrivenScannableMotionUnitsTest, that means that in each
test method, the dummyMotor object starts at position 0.

If DummyMotor is changed to save its position, testAsynchronousMoveTo does a
move from 1.0 to 1.0 and assumes it will take at least 50 ms. This is unlikely
to pass.

1. Make testAsynchronousMoveTo move the scannable to 0 before moving to 1.0.
   Then we know the move to 1.0 should take about one second.

2. Replace 2-element positions with 1-element positions, as the underlying
   scannable has one dimension.

3. Add some explanation as to how this thing works!

Change-Id: I7110aaac79d47deb2710e2bd86e45bf501aecb4d