Xspress3 v2 testing Xspress3 is still under development in both GDA and EPICs.
Testing Xspress3 v2 currently and need some updates in GDA to cope the changes 
(new PVS and new behavior). Due to the limited time to have GDA 9 working on I18 and better 
Xspress3 performance . The following was needed:
- Keep the Xspress3 interface, pvProvider and Controller the same as adding new CA to 
PV will not affect the currently running version as it uses LazyPV class and the 
connection to EPICs PV is only done when requested.
- Create a v2 version for the remaining classes as big changes were done:
          to improve the scan performance and reliability. 
          to take into account the new EPICs PV 
          the new HDF5 file library and the lack of performance of 
          NexusExtractor to read the file.

That is a working version with EPICs v2 tested with hardware on I18. Further 
changes will be needed after EPICs v2 will stabilized.

Conflicts:
	uk.ac.gda.devices.xspress3/META-INF/MANIFEST.MF

Change-Id: I645aac8614e156596e5c561d2be1b2a69acd04e3