Add property to enable SWMR mode for our nexus scan files.
The property is gda.nexus.writeSwmr

This will cause the NexusDataWriter to activate swmr mode just before
the first scan point is written to file, but after the initial structure
of the file is created. This structure cannot be changed once in swmr
mode (i.e. new datasets and groups may not be created).

This also forces the files to be written with the new binary format and
so should not be used in production until the file format is fixed by
the HDF group (i.e. not subject to change)

Change-Id: Ie3c6b5304fa30ad8584beab3dddec59a040d2ef2