GDA-6243 Logpanel: don't filter out log entries with stack traces
Initially the Pattern used by MatchingFilter to filter out log entries is null,
which causes all log entries to be shown.

If filter text is entered, a Pattern is built using a regex of the form
".*TEXT.*".

If the filter text box is then cleared, the regex is set to ".*.*".

The text displayed for a log entry with a stack trace includes newlines, but by
default '.' in Pattern regexes doesn't match newlines.

This meant that if filter text was entered, log entries with stack traces would
not be shown, even if the filter text was found somewhere in the log entry.

If filter text was entered and then cleared, no log entries with stack traces
would be displayed at all!

Changes:

1. Specify the DOTALL flag when building the Pattern; this causes log entries
   with stack traces to match the entered text.

2. Set the Pattern to null if the filter text box is cleared. This is more
   efficient than setting it to match ".*.*".

Change-Id: I389531504d41503a68a7079722aa850c12c88c9f