Add XAS feature and xspress3 visibility and remove circular dependency
The new XAS feature references all the plugins to support i18 and the
xspress3.epics plugin is made visibile (in osgi) to Spring, CORBA/RMI 
and Jython.

Also the reference to uk.ac.gda.devices.detector.xspress3.controllerimpl 
in the xspress3 Manifest causes a circular reference and is not used in 
the plugin so it has been removed.
Change-Id: I7a39024654e8502a2379d46c04035796343f28a9