Move IExperimentObject to gda.api for Expt. Def. Queue in GDA9 server
Moving this allows aclient dependency to be disengaged from the 
experiment objects that get put on the queue. Also update affected 
manifests and plugin.

Also ExampleExperimentObject has a dummy, uncalled implementation of the interface 
in it to allow the checkin to proceed, this will be replaced in the 
next checkin.

Change-Id: I0fc027fa4ae07f4618e43752aca8bb74bba08318