Move ScanObject to allow Experiment queue to work in GDA9 server
The ScanObject and other implementors of IExperimentObject need to be 
available on the server with minimal dependencies to allow it to get 
them off the experiment queue. This change also removes the extra 
method from IExperiment object introduced in change 1849 and allows the 
server to read the scan type from the copy of the preferences now set 
by the client instead rather than directly from the client itself.

Change-Id: Iae279ee3e06a304246d53f4410863f478fe32f57