Fix IllegalAccessError on file save
Because of the packages of the sub and base classes of the data writer 
class(es) being split across different packages and the fact that they 
were making use of protected and package private calls, the OSGI 
classloader did not have access to some methods/field of the base class 
(IncrementalFile). This resulted in an IllegalAccessError. To fix this, 
the affected methods are now public and the fields are accessible via 
getters.

Change-Id: I3aa1a2180cb22b07d47bcf7d9ce85a670bc2c74b