Remove need to access Spring via Ext classloader in RMI exporting
Up until now, when exporting services for RMI, the server has had to 
rely on the ext classloader pointing to the required Spring jars in a 
specially created directory. This is undesirable and makes the server 
product build job unnecessarily complex. To correct this, the 
GDARmiServiceExporter class now makes use of the thread Context 
Classloader, setting it to it's own classloader, which allows it to see 
the Spring jars in their normal location, since the one provided by OSGi 
is unable to in this situation. This requires the bundle containing the 
service interface to be an Eclipse buddy of uk.ac.gda.core (where the 
GDARmiExporter class lives) which means it must also depend on 
uk.ac.gda.core.



Change-Id: Ic0929e1ba0e93ebdc7fc668a54b30b2bdaa3f5f7