Disconnect workspace location from main application root
Allow use of the -data option to specify the server workspace such that 
its location can be independent of the install location of the server 
(parent of workspace when running in Eclipse, anywhere when an exported 
product). This allows user specific workspace to be used when the 
server is centrally installed on dls_sw.

Also add --launcher.suppressErrors to the product launcher to stop it 
popping up unwanted jvm exit message windows when shutting down.

Change-Id: I72266c5fd27bf712dad57d8242d456718777de74