Change NexusGroupData to automatically set chunking.
Changes NexusGroupData to estimate a chunk size in its constructors.

Chunking is not estimated for string datasets.

Estimation prioritises the last dimensions (e.g. 1 x 128 x 2048 is
preferred over 32 x 64 x 128 - both are 1MB with 4 byte data, but the
first is more likely to be suitable for writing out detector data).