Miscellaneous fixes to DummyDAServer
(1) Reduce the level of the "DAServer command" message from info to
debug as it can potentially be generated very frequently
- see for example HardwareTriggeredPCO.prepareForCollection(),
which calls it in a 50 ms loop via Etfg.getStatus()

(2) Clear the stopRun flag when starting a run.

(3) When deciding when to pause, check for livePause > 0 rather 
than just livePause == 1.

Change-Id: I57f6d39b3ba8c087d0ed567a1445dff9f2ed0da5