Restore line in ZebraImpl.pcArm() that disappeared in merge
A line in ZebraImpl.pcArm():
	pvFactory.getPVInteger(PCArm).putWait(1);

which is the only statement that actually does anything, was removed
when an 8.44 branch was merged in. I suspect it was something to do
with the fact that dev was renamed to pvFactory in 8.52, which Git
regards as a delete and add, and the delete was applied in the merge
but not the add.

Change-Id: I89741f9e6830870f5e48aec37a9959c4482111a9