DATAACQTEAM-276 Prevent IDfind.ext errors in editor windows in e4 client
When a new editor view opens, e4 expects the 
IWorkbenchActionConstants.FIND_EXT to be defined in order to be able to 
update the menus when contributors are used. This leads to "Cannot find 
IDfind.ext" errors when affected editor views are opened. This is 
actually a bad concatenation and should read "Cannot find ID find.ext" 
indicating that the required id has not been initialised. Adding such an 
initialiser to ApplicationActionBarAdvisor prevents these errors, fixes 
the PyDev editor and allows the contributors previously removed from the 
i18 custom editors to be reinstated. 

Change-Id: Id629b7427d8a961ac180678c578e9b57073cf208