Fix ScannableGroup to continue to format output if a device fails
Previously, if ScannableGroup.toFormattedString() encountered an
exception in formatting one device, it would just throw an exception
and stop, so none of the devices in the group would be displayed.

This changes causes it to print "n/a" in place of a value it cannot get
and continue with the rest of the group.

Change-Id: I2232e4ce0e8a2e9e6b5d95af73deb854df510ddf