I18-44 Modify ConcurrentScan to control whether update messages are sent to GUI
Add a property sendUpdateEvents to ConcurrentScan to control whether
EventType.UPDATED events are sent during scanning.

This has been added to handle a particular issue with some scans
(notably ConstantVelocityScanLine), where update messages from "dummy"
pre-scans can swamp the GUI.

Scans for which an issue can set this property false to suppress
these messages.

This is not a particularly nice solution, but the problem will
disappear when the new scanning mechanism is active.

Change-Id: I1b8f6d4b61b19c19eab95dd278c1078ce872a374