DAQ-134 Update launcher to a working configuration
This isn't a proper fix, really the dependencies in the feature need
updating so the product will actually be valid, but having a working
launcher might be of help to someone trying to fix the dependencies.

Of course, if dependencies elsewhere in the system change, this launcher
might stop working anyway. But at least it's a bit closer to working
than it would have been without this change!

Change-Id: I8cfc98c9192937eb89c89da1bbdbd57353268785