Update MappingScanNewStyleEventObserver to not use the map bean
With the ability to get the NeXus file tree remotely this bean is not
needed anymore only the file path is required.

Also change to using the new EventConstants interface to refer to the
STATUS_TOPIC constant, to remove a warning.

Cleaned up the properties used for the dataserver

Refactored file open and close constants

Updated comments and logging to make clearer what is happening at the
end of the scan

Change-Id: Idf41b8064bf1b44933468959db1d6aa767f6a4e2