DAQ-104 Add scisoft plotclient to allow Jython plotting functionality
The Jython terminal plotting functionality has not been available in
GDA9 server so far since the scisoft plotclient bundle was not present.
This was because it had dependencies on lots of client/ui library code
via a dependency on org.dawb.common.ui. This has now been removed
allowing the plotclient bundle to be added without affecting the 
server's dependency tree. This in turn will allow changes to support
class loading from the plotclient bundle.

Change-Id: Iaa36ac42b1ce93f1deaf56646bd604c34a9a0833