Tidy up and document LocalProperties.getAsInt() to be less confusing
Also add a couple more unit tests to verify behaviour

All uses of the methods in an all-dls-config GDA workspace were checked,
and none relied on using the object Integer type or checked for a null
value before unboxing to an int. This change should therefore not make
any difference to current behaviour in any actual usage, but hopefully
the purpose and results of the methods will be better defined.

Change-Id: I836ce33ccc01a34b3fa16cb061b327ddc4472331