Stop the proliferation of beamline specific server launchers in GDA 9
As more beamlines come on stream with GDA 9 we could end up with many,
more slightly out of step with the correct version, beamline specific 
launchers that don't have up to date arguments/plugin list etc. To stop
this we need to control them centrally via the Core team so that the
only checked in launchers are ones that correctly reflect the current
state of the server product.

This commit removes the launchers that have already built up and
replaces them with dummy mode and live mode prototype launchers derived
from the server product feature. Along with these is a set of 
instructions on using them to produce local beamline specific launchers
that won't end up then getting checked back into git each time they are
modified.

Change-Id: I87f68b2e61154e4deddc2d1091adb777e2b1f64f