DAQ-104 Rearrange visibility extension point for better readability
Currently visible packages, the bundles they belong to and script source
folders are all mixed up in one list which makes them difficult to
separate in the Extension editor as the package list inevitably
dominates. This splits things up as follows:

1. Separate the script source locations into their own extension point
2. Make the package list a separate sequence within the main sequence
so that if the optional bundle element is present, it does not get lost
amongst all the packages.

Change-Id: Ida35144d8f317bbcc0600ead2ca14054aa34efd4