I21-24 I21-39 Added AppleII ID controller
Control is split into two parts:

- Device classes (Epics & dummy) implementing IApple2ID and based on
Apple2IDBase which take a set of gap & phase motor positions and
move the ID, if it is safe to do so.

- Scannable classes allowing the user to specify the desired ID position
in terms of gap & polarisation and translates these into motor positions.
For the moment, two classes have been added:

- Apple2IDGapPolarPos which allows all four phases supported by the
ID (LH, C, L1, L2) and allows the user to specify the position of the phase motors.

- Apple2IDGapPolar which allows two phases: horizontal (LH) and
vertical (LV), where LV is implemented as circular with maximum motor
position.

These Scannable classes are provisional. In future, there will at least
be a requirement for scanning energy.

Change-Id: I5fb86790133a77f3f1062ce72c358deb4c939928