I15_1-111 (9.1pre) Remove dependency on gda and config symbolic links.
The config link was deprecated in the standard GDA config. It should
only be present as a shortcut for developers navigating the filesystem,
and no scripts, links or configuration paths should ever expect it to be
present.

Also, with the standard config, scripts and configuration paths should
avoid pointing outside of GDA_WORKSPACE_PARENT, since this could cause
a mismatch between the expected configuration and what is actually used.

Change-Id: Ia79eb583b53d6e6c79e8c54187fec9b7a02837e5