Xspress3Detector: move calculation of how many frames to collect to a new protected method
MX fluorescence scans are done using a MultiRegionScan. That type of scan can't
return the dimensions of the scan, so Xspress3Detector can't use the
ScanInformation to figure out how many frames to collect.

Moving that code to a protected method allows us to override it for MX and fix
the number of frames to one.

Change-Id: I1f96a787da49f0b9aea3fe18232d7e2d9d037773