Xspress3Detector: don't set number of frames in atScanStart if a MultiRegionScan is running
This is an improvement to a69f8d7.

For MX fluo scans, we have a custom Detector object that delegates to the
Xspress3Detector object, so we already have a place in the code to force the
number of frames to 1.

However we don't want the number of frames set to -1 by Xspress3Detector due to
it being used in a MultiRegionScan.

So rather than subclassing Xspress3Detector and overriding
calculateNumberOfFramesToCollect, we just need a way to prevent the number of
frames being set to -1 (which is this change), and then we can force it to 1 in
our custom Detector object.

Change-Id: I3f24eff4977ca0e19fd6cd640467fcf402fa8a57