JythonTerminalView: change terminal_output_*.txt filename format to use timestamp instead of number
Up to now these output files were named terminal_output_XXX.txt where XXX was a
number. When a new file is to be created, the highest used number is found and
the new file uses the next number.

We are encouraged to move log files out of /dls_sw regularly, but if we move
old terminal_output_*.txt files, the old numbers will be reused. This means new
output files will clash with old ones when we archive the newer ones.

By using a timestamp in the filenames, the files will always have unique names.

Change-Id: I08c3bf4a4746d6c89de7d3bbb34fb44c8c7ac539