I15_1-111 (9.1) Don't clobber requested launcher when using devel latch.
If you try to start the servers from a non beamline deployment using a
development latch to point back to that deployment, then the script
ignores the latch, and instead uses the latch of the currently active
beamline deployment to determine which gda server to run. Without this
fix, you would not be able to run any version of the GDA server which
wasn't currently active on the beamline, making offline testing
impossible.

Change-Id: I8fadd7bbb656ac9785c5ad3a6d2b024c8ef202a0